Move IsAssetFilterTrait out of MetaData namespace.
